btrfs: index free space cache by size as well
Summary:
btrfs: index free space cache by size as well

A long time ago we indexed free space based on size.  This was removed
because we thought the overhead was unhelpful.  However in highly
fragmented file systems the overhead of searching for the best fit is
much higher than the overhead of indexing every entry on offset and
size.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: Josef Bacik <josef@toxicpanda.com>