posix-cpu-timers: Replace cpu_timer_pid_type with clock_pid_type
Taking a clock and returning a pid_type is a more general and
a superset of taking a timer and returning a pid_type.

Perform this generalization so that future changes may use
this code on clocks as well as timers.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
(cherry picked from commit fece98260f31292dd468925c7582065bf6193212)
Signed-off-by: Jonathan Lemon <jonathan.lemon@gmail.com>